refactor: topics.csv -> topics.txt
The CSV file format is incorrectly used for the topics file,
since the topics file doesn't escape commas in its normal use.

Not only does this render the CSV extension useless
(e.g. in github rich renders), but it also may be a bit misleading
on its file name as well.